requirements.txt
python-telegram-bot>=21.10
python-dotenv>=1.0.1
Flask~=3.1.0
pyTelegramBotAPI~=4.26.0
Quart~=0.20.0
Hypercorn~=0.17.3